add complete algorithm prototype

+++ world/src/main/scala/com/sberhack/fun/car/Car.scala
@@ -8,7 +8,7 @@ case class Car(
                 name: String,
                 cash: Double,
                 cashLimit: Double,
-                currentNodeId: Long,
+                currentNodeId: Int,
                 nextMove: Option[CarNextMove]
               ) {
