+++ world/src/main/scala/com/sberhack/fun/graph/node/VaultData.scala
@@ -3,7 +3,7 @@ package com.sberhack.fun.graph.node
 import com.sberhack.fun.car.Car
 
 case class VaultData (
-                       id: Long,
+                       id: Int,
                        money: Double, // Количество денег которые уже отвезены в хранилище
                        cars: Seq[Car]
                      )
